requirements.txt
fastapi==0.110.1
pymongo==4.6.3
uvicorn==0.29.0
pydantic~=2.7.0
bcrypt~=4.1.2
PyJWT~=2.8.0
python-multipart
python-dotenv~=1.0.1